Still working

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/TransferModel.java
@@ -3,51 +3,50 @@ package com.techelevator.tenmo.model;
 import java.math.BigDecimal;
 
 public class TransferModel {
-    private int transfer_id;
-    private int transfer_type_id;
-    private int transfer_status_id;
-    private int account_from;
-    private int account_to;
+    private long transfer_id;
+    private long transfer_type_id;
+    private long transfer_status_id;
+    private long account_from;
+    private long account_to;
     private BigDecimal amount;
 
-
-    public int getTransfer_id() {
+    public long getTransfer_id() {
         return transfer_id;
     }
 
-    public void setTransfer_id(int transfer_id) {
+    public void setTransfer_id(long transfer_id) {
         this.transfer_id = transfer_id;
     }
 
-    public int getTransfer_type_id() {
+    public long getTransfer_type_id() {
         return transfer_type_id;
     }
 
-    public void setTransfer_type_id(int transfer_type_id) {
+    public void setTransfer_type_id(long transfer_type_id) {
         this.transfer_type_id = transfer_type_id;
     }
 
-    public int getTransfer_status_id() {
+    public long getTransfer_status_id() {
         return transfer_status_id;
     }
 
-    public void setTransfer_status_id(int transfer_status_id) {
+    public void setTransfer_status_id(long transfer_status_id) {
         this.transfer_status_id = transfer_status_id;
     }
 
-    public int getAccount_from() {
+    public long getAccount_from() {
         return account_from;
     }
 
-    public void setAccount_from(int account_from) {
+    public void setAccount_from(long account_from) {
         this.account_from = account_from;
     }
 
-    public int getAccount_to() {
+    public long getAccount_to() {
         return account_to;
     }
 
-    public void setAccount_to(int account_to) {
+    public void setAccount_to(long account_to) {
         this.account_to = account_to;
     }
 
@@ -58,5 +57,4 @@ public class TransferModel {
     public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
-
 }
